+++ src/libcsg/cgengine.cc
@@ -17,6 +17,7 @@
 
 // Standard includes
 #include <fstream>
+#include <memory>
 
 // VOTCA includes
 #include <votca/tools/tokenizer.h>
@@ -34,19 +35,13 @@ namespace po = boost::program_options;
 
 CGEngine::CGEngine() = default;
 
-CGEngine::~CGEngine() {
-  for (auto &_molecule_def : _molecule_defs) {
-    delete _molecule_def.second;
-  }
-  _molecule_defs.clear();
-}
-
 /**
     \todo melts with different molecules
 */
-TopologyMap *CGEngine::CreateCGTopology(Topology &in, Topology &out) {
+std::unique_ptr<TopologyMap> CGEngine::CreateCGTopology(Topology &in,
+                                                        Topology &out) {
   MoleculeContainer &mols = in.Molecules();
-  TopologyMap *m = new TopologyMap(&in, &out);
+  auto m = std::make_unique<TopologyMap>(&in, &out);
   for (const auto &mol : mols) {
     if (IsIgnored(mol->getName())) {
       continue;
@@ -77,11 +72,11 @@ void CGEngine::LoadMoleculeType(string filename) {
 
   for (tools::Tokenizer::iterator iter = tok.begin(); iter != tok.end();
        ++iter) {
-    CGMoleculeDef *def = new CGMoleculeDef();
+    auto def = std::make_unique<CGMoleculeDef>();
     string file = *iter;
     boost::trim(file);
     def->Load(file);
-    _molecule_defs[def->getIdent()] = def;
+    _molecule_defs[def->getIdent()] = std::move(def);
   }
 }